Delete Object & Create Account Updates
Create Account:
- Role = admin requires user to enter the properties file defined passcode or the user is downgraded to role = user

- All Object deletion requires the user to be a validated administrator.

+++ MSSE670_CourseProject_FoodMenu/src/com/foodmenu/model/business/managers/DayMenuManager.java
@@ -4,8 +4,10 @@ import java.util.ArrayList;
 import java.util.Calendar;
 
 import com.foodmenu.model.business.exceptions.ServiceLoadException;
+import com.foodmenu.model.business.exceptions.UserPrivilegesException;
 import com.foodmenu.model.business.factory.ServiceFactory;
 import com.foodmenu.model.domain.DayMenu;
+import com.foodmenu.model.domain.User;
 import com.foodmenu.model.services.daymenuservice.IDayMenuService;
 import com.foodmenu.model.services.exceptions.DayMenuServiceException;
 import com.foodmenu.model.services.exceptions.FoodItemServiceException;
@@ -17,7 +19,13 @@ import com.foodmenu.model.services.exceptions.MenuItemServiceException;
  */
 public class DayMenuManager {
 
+	private User user;
+	
 	public DayMenuManager() {
+	}
+	
+	public DayMenuManager(User user) {
+		this.user = user;
 	}
 	
 	/** 
@@ -39,16 +47,21 @@ public class DayMenuManager {
 	/** 
 	 * Use Case : DayMenu-310
 	 * Delete Existing Day Menu
+	 * @throws UserPrivilegesException 
 	 */
 	public boolean deleteDayMenu(DayMenu dayMenu) throws ServiceLoadException, 
-		DayMenuServiceException {
+		DayMenuServiceException, UserPrivilegesException {
 		
-		ServiceFactory serviceFactory = new ServiceFactory();
-		IDayMenuService dayMenuSvc = (IDayMenuService)serviceFactory.getService("IDayMenuService");
-		if(dayMenuSvc.deleteDayMenuData(dayMenu)) {
-			return true;
+		if(this.user.getRole().equals("admin")) {
+			ServiceFactory serviceFactory = new ServiceFactory();
+			IDayMenuService dayMenuSvc = (IDayMenuService)serviceFactory.getService("IDayMenuService");
+			if(dayMenuSvc.deleteDayMenuData(dayMenu)) {
+				return true;
+			} else {
+				return false;
+			}
 		} else {
-			return false;
+			throw new UserPrivilegesException(String.format("User %s isn't an admin, and therefore does not have the \nappropriate privileges to perform delete task!", user.getEmailAddress()));
 		}
 	}
